+++ server/client/pages/DFH/SmartTrade/Handler/MockHandler.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function MockHandler({ signer, routerAdapter, adapters, searchParams }) {
+  return <div>Mock handler</div>;
+}
